+++ TinyNvidiaUpdateChecker/MainConsole.cs
@@ -81,6 +81,11 @@ namespace TinyNvidiaUpdateChecker
         public static bool showUI = true;
 
         /// <summary>
+        /// Disable "Press any key to exit..." prompt
+        /// </summary>
+	public static bool noPrompt = false;
+
+        /// <summary>
         /// Enable extended information
         /// </summary>
         public static bool debug = false;
@@ -239,11 +244,9 @@ namespace TinyNvidiaUpdateChecker
                 }
             }
 
-            Console.WriteLine();
-            Console.WriteLine("Press any key to exit...");
-            if (showUI) Console.ReadKey();
             LogManager.Log("BYE!", LogManager.Level.INFO);
-            Environment.Exit(0);
+            callExit(0);
+	
         }
 
         /// <summary>
@@ -308,6 +311,10 @@ namespace TinyNvidiaUpdateChecker
                     FreeConsole();
                     showUI = false;
                 }
+				
+		else if (arg.ToLower() == "--noprompt") {
+                    noPrompt = true;
+                }
 
                 // erase config
                 else if (arg.ToLower() == "--erase-config") {
@@ -361,6 +368,7 @@ namespace TinyNvidiaUpdateChecker
                     Console.WriteLine($"Usage: {Path.GetFileName(Environment.ProcessPath)} [ARGS]");
                     Console.WriteLine();
                     Console.WriteLine("--quiet                      Runs the application quietly in the background, and will only notify the user if an update is available.");
+                    Console.WriteLine("--noprompt                   Runs the application without prompting to exit.");
                     Console.WriteLine("--erase-config               Erase configuration file.");
                     Console.WriteLine("--debug                      Turn debugging on, will output more information that can be used for debugging.");
                     Console.WriteLine("--force-dl                   Force prompt to download drivers, even if the user is up-to-date - should only be used for debugging.");
@@ -429,10 +437,7 @@ namespace TinyNvidiaUpdateChecker
                 Console.WriteLine();
                 // todo use pcilookup API https://www.pcilookup.com/api.php?action=search&vendor=10DE&device=13C2
                 Console.WriteLine("No supported NVIDIA GPU was found! If you have a NVIDIA GPU then manually install a driver for it first, then use TNUC to keep it updated.");
-                Console.WriteLine();
-                Console.WriteLine("Press any key to exit...");
-                if (showUI) Console.ReadKey();
-                Environment.Exit(1);
+                callExit(1);		    
             }
 
             return (gpuName, isNotebook);
@@ -468,20 +473,14 @@ namespace TinyNvidiaUpdateChecker
                 Console.WriteLine();
                 Console.WriteLine($"gpuName:    {gpuName}");
                 Console.WriteLine($"isNotebook: {isNotebook}");
-                Console.WriteLine();
-                Console.WriteLine("Press any key to exit...");
-                if (showUI) Console.ReadKey();
-                Environment.Exit(1);
+                callExit(1);
             } catch (Exception ex) {
                 Console.Write("ERROR!");
                 Console.WriteLine();
                 Console.WriteLine("Unable to retrieve GPU data. Do you have working internet connectivity?");
                 Console.WriteLine();
                 Console.WriteLine(ex.ToString());
-                Console.WriteLine();
-                Console.WriteLine("Press any key to exit...");
-                if (showUI) Console.ReadKey();
-                Environment.Exit(1);
+                callExit(1);
             }
 
             // Get operating system ID
@@ -506,10 +505,7 @@ namespace TinyNvidiaUpdateChecker
                 Console.WriteLine("Unable to retrieve OS data.");
                 Console.WriteLine();
                 Console.WriteLine(ex.ToString());
-                Console.WriteLine();
-                Console.WriteLine("Press any key to exit...");
-                if (showUI) Console.ReadKey();
-                Environment.Exit(1);
+                callExit(1);
             }
 
             foreach (var os in osData) {
@@ -527,10 +523,7 @@ namespace TinyNvidiaUpdateChecker
                 Console.WriteLine($"gpuName:   {gpuName}");
                 Console.WriteLine($"osVersion: {osVersion}");
                 Console.WriteLine($"osBit:     {osBit}");
-                Console.WriteLine();
-                Console.WriteLine("Press any key to exit...");
-                if (showUI) Console.ReadKey();
-                Environment.Exit(1);
+                callExit(1);                    
             }
 
             // Check for DCH for newer drivers
@@ -583,11 +576,8 @@ namespace TinyNvidiaUpdateChecker
                 Console.WriteLine();
                 Console.WriteLine(ex.ToString());
             }
-
-            Console.WriteLine();
-            Console.WriteLine("Press any key to exit...");
-            if (showUI) Console.ReadKey();
-            Environment.Exit(1);
+	
+            callExit(1);
 
             return null;
         }
@@ -618,10 +608,7 @@ namespace TinyNvidiaUpdateChecker
                 Console.Write("ERROR!");
                 Console.WriteLine();
                 Console.WriteLine("You are not connected to the internet!");
-                Console.WriteLine();
-                Console.WriteLine("Press any key to exit...");
-                if (showUI) Console.ReadKey();
-                Environment.Exit(2);
+                callExit(2);
             }
 
             if (SettingManager.ReadSettingBool("Minimal install")) {
@@ -953,6 +940,22 @@ namespace TinyNvidiaUpdateChecker
             }
         }
 
+        /// <summary>
+        /// Check for passed argument and prompt for exit if applicable
+        /// </summary>
+        /// 
+        private static void callExit(int exitNum)
+        {
+            if (!noPrompt)
+            {
+                Console.WriteLine();
+                Console.WriteLine("Press any key to exit...");
+            }
+		
+            if (showUI & !noPrompt) Console.ReadKey();
+            Environment.Exit(exitNum);
+        }	    
+	    
         private static bool DoesDriverFileSizeMatch(string absoluteFilePath) {
             return new FileInfo(absoluteFilePath).Length == downloadFileSize;
         }